Make integer narrowing explicit and enable implicit conversion checking
Cast intentional narrowing into byte and 16-bit storage at 518 sites so
clang-diagnostic-implicit-int-conversion and its negation companion can be
enforced. Along the way: return the full decompressed size from both games'
Load_Uncompress instead of the truncated 16-bit header field, assign the RA
foot speed directly rather than through a byte reference, widen TD's
Get_Ownable virtual to cover all ten houses, and derive the null-modem
connect ID with std::bit_cast.

+++ src/ra/2keyfram.cc
@@ -242,10 +242,10 @@ void* Build_Frame(const void* dataptr, const uint16_t framenumber,
     if (keyfr->x != kUncompressMagicNumber) {
       keyfr->x = kUncompressMagicNumber;
       if (IsTheaterShape) {
-        keyfr->y = TheaterSlotsUsed;
+        keyfr->y = static_cast<uint16_t>(TheaterSlotsUsed);
         TheaterSlotsUsed++;
       } else {
-        keyfr->y = TotalSlotsUsed;
+        keyfr->y = static_cast<uint16_t>(TotalSlotsUsed);
         TotalSlotsUsed++;
       }
       /*
