+++ src/ra/fly.cc
@@ -82,7 +82,7 @@ ImpactType FlyClass::Physics(COORDINATE& coord, DirType facing) {
     */
     if (result.quot) {
       COORDINATE newcoord;  // New working coordinate.
-      newcoord = Coord_Move(coord, facing, actual);
+      newcoord = Coord_Move(coord, facing, static_cast<unsigned short>(actual));
       /*
       **	If no movement occurred, then presume it hasn't moved at all
       **	and return immediately with this indication.
