+++ src/ra/sendfile.cc
@@ -700,8 +700,9 @@ bool Send_Remote_File(char* file_name, int gametype) {
       if (block_number < total_blocks) {
         if (NullModem.Num_Send() < 2) {
           send_packet.Command = SERIAL_FILE_CHUNK;
-          send_packet.BlockNumber = block_number;
-          send_packet.BlockLength = std::min(file_length, max_chunk_size);
+          send_packet.BlockNumber = static_cast<unsigned short>(block_number);
+          send_packet.BlockLength = static_cast<unsigned short>(
+              std::min(file_length, max_chunk_size));
 
           file_length -= send_packet.BlockLength;
 
@@ -736,8 +737,9 @@ bool Send_Remote_File(char* file_name, int gametype) {
       if (block_number < total_blocks) {
         if (Ipx.Global_Num_Send() == 0) {
           send_packet.Command = SERIAL_FILE_CHUNK;
-          send_packet.BlockNumber = block_number;
-          send_packet.BlockLength = std::min(file_length, max_chunk_size);
+          send_packet.BlockNumber = static_cast<unsigned short>(block_number);
+          send_packet.BlockLength = static_cast<unsigned short>(
+              std::min(file_length, max_chunk_size));
 
           file_length -= send_packet.BlockLength;
